backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
httpx==0.26.0
pydantic==2.5.3
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
